웹메인과 콘솔메인의 형태를 비슷하게 하였습니다.
초기에 만들어진 가정이 달랐어서 구조가 조금 상이합니다.
(웹 기준으로 한화면에 모든 인풋을 받을거라고 생각했었습니다.)
